rename everything to store_redirect
