+++ android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'launch_review'
+rootProject.name = 'store_redirect'
