+++ android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.iyaffle.launchreview">
+  package="com.danieldallos.storeredirect">
 </manifest>
