--- ios/Classes/LaunchReviewPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface LaunchReviewPlugin : NSObject<FlutterPlugin>
-@end
