+++ ios/Classes/StoreRedirectPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface StoreRedirectPlugin : NSObject<FlutterPlugin>
+@end
